refactor(core): move metaserver_worker_loop into storage.rs
StorageEngine now owns full metaserver lifecycle: channel setup, worker
spawn, and send. PegaEngine.set_metaserver_client delegates to StorageEngine.
